Fix #8199: Adds 16:9 aspect ratio cropping for topic thumbnail (#8303)
* initial changes

* fix lint issues

* fix lint issues

* use webpack to deliver cropper

* fix typescript test
