+++ package.json
@@ -56,6 +56,7 @@
     "cache-loader": "^4.1.0",
     "clean-webpack-plugin": "^2.0.1",
     "core-js": "^2.5.3",
+    "cropperjs": "^1.5.6",
     "css-loader": "^3.1.0",
     "d3": "^5.9.2",
     "dotenv": "^8.1.0",
