+++ webpack.common.config.ts
@@ -671,6 +671,7 @@ module.exports = {
       test: /\.css$/,
       include: [
         path.resolve(__dirname, 'extensions'),
+        path.resolve(__dirname, 'node_modules'),
       ],
       use: ['style-loader', 'css-loader']
     }]
